requirements.txt
pytorch_lightning==2.1.3
omegaconf==2.3.0
jaxtyping
typeguard
nerfacc==0.5.3
#git+https://github.com/NVlabs/tiny-cuda-nn/#subdirectory=bindings/torch
diffusers==0.25.0
transformers==4.28.1
accelerate
opencv_python
matplotlib
imageio>=2.28.0
imageio[ffmpeg]
#git+https://github.com/NVlabs/nvdiffrast.git
libigl
xatlas
trimesh[easy]
networkx 
pysdf
tensorboard
wandb
gradio
envlight==0.1.0
torchmetrics==1.2.1
huggingface_hub==0.24.6
numpy==1.26.2

# for zero123
clip
einops
kornia
taming-transformers-rom1504

#controlnet
controlnet_aux
